Finalizando Formação de java-spring-web completo

+++ src/main/java/br/com/alura/screenmatch/model/DadosTemporada.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 
 @JsonIgnoreProperties(ignoreUnknown = true)
-public record DadosTemporada(@JsonAlias("Season") Long numero,
+public record DadosTemporada(@JsonAlias("Season") Integer numero,
                              @JsonAlias("Episodes")List<DadosEpisodio> episodios ) {
