+++ src/main/java/br/com/alura/screenmatch/repository/SerieRepository.java
@@ -36,6 +36,6 @@ public interface SerieRepository extends JpaRepository<Serie, Long> {
     List<Serie> lancamentosMaisRecentes();
 
     @Query("SELECT e FROM Serie s JOIN s.episodios e WHERE s.id = :id AND e.temporada = :numero")
-    List<Episodio> obterEpisodiosPorTemporada(Long id, Long numero);
+    List<Episodio> obterEpisodiosPorTemporada(Long id, Integer numero);
 
 }
